login form with menu, change default images.

+++ app/Http/Controllers/BannerCategoryController.php
@@ -39,7 +39,7 @@ class BannerCategoryController extends Controller
         if($banner->category_id == 1) {
             if (isset($request->photo1_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo1_url))){
+                if(file_exists(public_path($banner->photo1_url)) and !empty($banner->photo1_url)){
                     unlink(public_path($banner->photo1_url));
                 }
                 $fileName = $request->photo1_url->store('public/principal_banner');
@@ -47,7 +47,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo2_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo2_url)) and !empty($banner->photo2_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo2_url->store('public/principal_banner');
@@ -55,7 +55,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo3_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo3_url)) and !empty($banner->photo3_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo3_url->store('public/principal_banner');
@@ -63,7 +63,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo4_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo4_url))){
+                if(file_exists(public_path($banner->photo4_url)) and !empty($banner->photo4_url)){
                     unlink(public_path($banner->photo4_url));
                 }
                 $fileName = $request->photo4_url->store('public/principal_banner');
@@ -71,7 +71,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo5_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo5_url))){
+                if(file_exists(public_path($banner->photo5_url)) and !empty($banner->photo5_url)){
                     unlink(public_path($banner->photo5_url));
                 }
                 $fileName = $request->photo5_url->store('public/principal_banner');
@@ -82,7 +82,7 @@ class BannerCategoryController extends Controller
         if($banner->category_id == 2) {
             if (isset($request->photo1_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo1_url))){
+                if(file_exists(public_path($banner->photo1_url)) and !empty($banner->photo1_url)){
                     unlink(public_path($banner->photo1_url));
                 }
                 $fileName = $request->photo1_url->store('public/preescolar_banner');
@@ -90,7 +90,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo2_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo2_url)) and !empty($banner->photo2_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo2_url->store('public/preescolar_banner');
@@ -98,7 +98,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo3_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo3_url))){
+                if(file_exists(public_path($banner->photo3_url)) and !empty($banner->photo3_url)){
                     unlink(public_path($banner->photo3_url));
                 }
                 $fileName = $request->photo3_url->store('public/preescolar_banner');
@@ -106,7 +106,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo4_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo4_url))){
+                if(file_exists(public_path($banner->photo4_url)) and !empty($banner->photo4_url)){
                     unlink(public_path($banner->photo4_url));
                 }
                 $fileName = $request->photo4_url->store('public/preescolar_banner');
@@ -114,7 +114,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo5_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo5_url))){
+                if(file_exists(public_path($banner->photo5_url)) and !empty($banner->photo5_url)){
                     unlink(public_path($banner->photo5_url));
                 }
                 $fileName = $request->photo5_url->store('public/preescolar_banner');
@@ -125,7 +125,7 @@ class BannerCategoryController extends Controller
         if($banner->category_id == 3) {
             if (isset($request->photo1_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo1_url))){
+                if(file_exists(public_path($banner->photo1_url)) and !empty($banner->photo1_url)){
                     unlink(public_path($banner->photo1_url));
                 }
                 $fileName = $request->photo1_url->store('public/escuelacolegio_banner');
@@ -133,7 +133,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo2_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo2_url)) and !empty($banner->photo2_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo2_url->store('public/escuelacolegio_banner');
@@ -141,7 +141,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo3_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo3_url))){
+                if(file_exists(public_path($banner->photo3_url)) and !empty($banner->photo3_url)){
                     unlink(public_path($banner->photo3_url));
                 }
                 $fileName = $request->photo3_url->store('public/escuelacolegio_banner');
@@ -149,7 +149,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo4_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo4_url))){
+                if(file_exists(public_path($banner->photo4_url)) and !empty($banner->photo4_url)){
                     unlink(public_path($banner->photo4_url));
                 }
                 $fileName = $request->photo4_url->store('public/escuelacolegio_banner');
@@ -157,7 +157,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo5_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo5_url))){
+                if(file_exists(public_path($banner->photo5_url)) and !empty($banner->photo5_url)){
                     unlink(public_path($banner->photo5_url));
                 }
                 $fileName = $request->photo5_url->store('public/escuelacolegio_banner');
@@ -168,7 +168,7 @@ class BannerCategoryController extends Controller
         if($banner->category_id == 4) {
             if (isset($request->photo1_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo1_url))){
+                if(file_exists(public_path($banner->photo1_url)) and !empty($banner->photo1_url)){
                     unlink(public_path($banner->photo1_url));
                 }
                 $fileName = $request->photo1_url->store('public/superior_banner');
@@ -176,7 +176,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo2_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo2_url)) and !empty($banner->photo2_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo2_url->store('public/superior_banner');
@@ -184,7 +184,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo3_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo3_url))){
+                if(file_exists(public_path($banner->photo3_url)) and !empty($banner->photo3_url)){
                     unlink(public_path($banner->photo3_url));
                 }
                 $fileName = $request->photo3_url->store('public/superior_banner');
@@ -192,7 +192,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo4_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo4_url))){
+                if(file_exists(public_path($banner->photo4_url)) and !empty($banner->photo4_url)){
                     unlink(public_path($banner->photo4_url));
                 }
                 $fileName = $request->photo4_url->store('public/superior_banner');
@@ -200,7 +200,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo5_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo5_url))){
+                if(file_exists(public_path($banner->photo5_url)) and !empty($banner->evento_bg_picture)){
                     unlink(public_path($banner->photo5_url));
                 }
                 $fileName = $request->photo5_url->store('public/superior_banner');
@@ -211,7 +211,7 @@ class BannerCategoryController extends Controller
         if($banner->category_id == 5) {
             if (isset($request->photo1_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo1_url))){
+                if(file_exists(public_path($banner->photo1_url)) and !empty($banner->photo1_url)){
                     unlink(public_path($banner->photo1_url));
                 }
                 $fileName = $request->photo1_url->store('public/posgrados_banner');
@@ -219,7 +219,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo2_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo2_url)) and !empty($banner->photo2_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo2_url->store('public/posgrados_banner');
@@ -227,7 +227,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo3_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo3_url))){
+                if(file_exists(public_path($banner->photo3_url)) and !empty($banner->photo3_url)){
                     unlink(public_path($banner->photo3_url));
                 }
                 $fileName = $request->photo3_url->store('public/posgrados_banner');
@@ -235,7 +235,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo4_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo4_url))){
+                if(file_exists(public_path($banner->photo4_url)) and !empty($banner->photo4_url)){
                     unlink(public_path($banner->photo4_url));
                 }
                 $fileName = $request->photo4_url->store('public/posgrados_banner');
@@ -243,7 +243,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo5_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo5_url))){
+                if(file_exists(public_path($banner->photo5_url)) and !empty($banner->photo5_url)){
                     unlink(public_path($banner->photo5_url));
                 }
                 $fileName = $request->photo5_url->store('public/posgrados_banner');
@@ -254,7 +254,7 @@ class BannerCategoryController extends Controller
         if($banner->category_id == 6) {
             if (isset($request->photo1_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo1_url))){
+                if(file_exists(public_path($banner->photo1_url)) and !empty($banner->photo1_url)){
                     unlink(public_path($banner->photo1_url));
                 }
                 $fileName = $request->photo1_url->store('public/cursos_banner');
@@ -262,7 +262,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo2_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo2_url)) and !empty($banner->photo2_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo2_url->store('public/cursos_banner');
@@ -270,7 +270,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo3_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo3_url))){
+                if(file_exists(public_path($banner->photo3_url)) and !empty($banner->photo3_url)){
                     unlink(public_path($banner->photo3_url));
                 }
                 $fileName = $request->photo3_url->store('public/cursos_banner');
@@ -278,7 +278,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo4_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo4_url))){
+                if(file_exists(public_path($banner->photo4_url)) and !empty($banner->photo4_url)){
                     unlink(public_path($banner->photo4_url));
                 }
                 $fileName = $request->photo4_url->store('public/cursos_banner');
@@ -286,7 +286,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo5_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo5_url))){
+                if(file_exists(public_path($banner->photo5_url)) and !empty($banner->photo5_url)){
                     unlink(public_path($banner->photo5_url));
                 }
                 $fileName = $request->photo5_url->store('public/cursos_banner');
@@ -297,7 +297,7 @@ class BannerCategoryController extends Controller
         if($banner->category_id == 7) {
             if (isset($request->photo1_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo1_url))){
+                if(file_exists(public_path($banner->photo1_url)) and !empty($banner->photo1_url)){
                     unlink(public_path($banner->photo1_url));
                 }
                 $fileName = $request->photo1_url->store('public/eventos_banner');
@@ -305,7 +305,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo2_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo2_url))){
+                if(file_exists(public_path($banner->photo2_url)) and !empty($banner->photo2_url)){
                     unlink(public_path($banner->photo2_url));
                 }
                 $fileName = $request->photo2_url->store('public/eventos_banner');
@@ -313,7 +313,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo3_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo3_url))){
+                if(file_exists(public_path($banner->photo3_url)) and !empty($banner->photo3_url)){
                     unlink(public_path($banner->photo3_url));
                 }
                 $fileName = $request->photo3_url->store('public/eventos_banner');
@@ -321,7 +321,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo4_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo4_url))){
+                if(file_exists(public_path($banner->photo4_url)) and !empty($banner->photo4_url)){
                     unlink(public_path($banner->photo4_url));
                 }
                 $fileName = $request->photo4_url->store('public/eventos_banner');
@@ -329,7 +329,7 @@ class BannerCategoryController extends Controller
             }
             if (isset($request->photo5_url)) {
                 //Delete old images if necessary
-                if(file_exists(public_path($banner->photo5_url))){
+                if(file_exists(public_path($banner->photo5_url)) and !empty($banner->photo5_url)){
                     unlink(public_path($banner->photo5_url));
                 }
                 $fileName = $request->photo5_url->store('public/eventos_banner');
