+++ app/Http/Controllers/CourseSeminarController.php
@@ -93,7 +93,7 @@ class CourseSeminarController extends Controller
 
         if(isset($request->documento_pdf1)) {
             //Delete old if necessary
-            if(file_exists(public_path($courseseminar->documento_pdf1))){
+            if(file_exists(public_path($courseseminar->documento_pdf1)) and !empty($courseseminar->documento_pdf1)){
                 unlink(public_path($courseseminar->documento_pdf1));
             }
             $fileName = $request->documento_pdf1->store('public/posgrade_pdf');
@@ -101,7 +101,7 @@ class CourseSeminarController extends Controller
         }
         if(isset($request->documento_pdf2)) {
             //Delete old if necessary
-            if(file_exists(public_path($courseseminar->documento_pdf2))){
+            if(file_exists(public_path($courseseminar->documento_pdf2)) and !empty($courseseminar->documento_pdf2)){
                 unlink(public_path($courseseminar->documento_pdf2));
             }
             $fileName = $request->documento_pdf2->store('public/posgrade_pdf');
@@ -109,7 +109,7 @@ class CourseSeminarController extends Controller
         }
         if(isset($request->documento_pdf3)) {
             //Delete old if necessary
-            if(file_exists(public_path($courseseminar->documento_pdf3))){
+            if(file_exists(public_path($courseseminar->documento_pdf3)) and !empty($courseseminar->documento_pdf3)){
                 unlink(public_path($courseseminar->documento_pdf3));
             }
             $fileName = $request->documento_pdf3->store('public/posgrade_pdf');
