+++ app/Http/Controllers/InstitutionController.php
@@ -217,64 +217,62 @@ class InstitutionController extends Controller
         ]);
 
         $input = $request->all();
-
         //Delete old images if necessary
-        if(isset($request->institution_bg_picture)) {
-            if(file_exists(public_path($institution->institution_bg_picture))){
+        if(isset($request->institution_bg_picture) and !empty($institution->institution_bg_picture)) {
+            if(file_exists(public_path($institution->institution_bg_picture)))
                 unlink(public_path($institution->institution_bg_picture));
-            }
         }
-        if(isset($request->institution_picture_1)) {
+        if(isset($request->institution_picture_1) and !empty($institution->institution_picture_1)) {
             if(file_exists(public_path($institution->institution_picture_1))){
                 unlink(public_path($institution->institution_picture_1));
             }
         }
-        if(isset($request->institution_picture_2)) {
+        if(isset($request->institution_picture_2) and !empty($institution->institution_picture_2)) {
             if(file_exists(public_path($institution->institution_picture_2))){
                 unlink(public_path($institution->institution_picture_2));
             }
         }
-        if(isset($request->institution_picture_3)) {
+        if(isset($request->institution_picture_3) and !empty($institution->institution_picture_3)) {
             if(file_exists(public_path($institution->institution_picture_3))){
                 unlink(public_path($institution->institution_picture_3));
             }
         }
-        if(isset($request->institution_picture_4)) {
+        if(isset($request->institution_picture_4) and !empty($institution->institution_picture_4)) {
             if(file_exists(public_path($institution->institution_picture_4))){
                 unlink(public_path($institution->institution_picture_4));
             }
         }
-        if(isset($request->institution_picture_5)) {
+        if(isset($request->institution_picture_5) and !empty($institution->institution_picture_5)) {
             if(file_exists(public_path($institution->institution_picture_5))){
                 unlink(public_path($institution->institution_picture_5));
             }
         }
-        if(isset($request->institution_picture_6)) {
+        if(isset($request->institution_picture_6) and !empty($institution->institution_picture_6)) {
             if(file_exists(public_path($institution->institution_picture_6))){
                 unlink(public_path($institution->institution_picture_6));
             }
         }
-        if(isset($request->banner_inst_picture_1)) {
+        if(isset($request->banner_inst_picture_1) and !empty($institution->banner_inst_picture_1)) {
             if(file_exists(public_path($institution->banner_inst_picture_1))){
                 unlink(public_path($institution->banner_inst_picture_1));
             }
         }
-        if(isset($request->banner_inst_picture_2)) {
+        if(isset($request->banner_inst_picture_2) and !empty($institution->banner_inst_picture_2)) {
             if(file_exists(public_path($institution->banner_inst_picture_2))){
                 unlink(public_path($institution->banner_inst_picture_2));
             }
         }
-        if(isset($request->banner_inst_picture_3)) {
+        if(isset($request->banner_inst_picture_3) and !empty($institution->banner_inst_picture_3)) {
             if(file_exists(public_path($institution->banner_inst_picture_3))){
                 unlink(public_path($institution->banner_inst_picture_3));
             }
         }
-        if(isset($request->banner_inst_picture_4)) {
+        if(isset($request->banner_inst_picture_4) and !empty($institution->banner_inst_picture_4)) {
             if(file_exists(public_path($institution->banner_inst_picture_4))){
                 unlink(public_path($institution->banner_inst_picture_4));
             }
         }
-        if(isset($request->banner_inst_picture_5)) {
+        if(isset($request->banner_inst_picture_5) and !empty($institution->banner_inst_picture_5)) {
             if(file_exists(public_path($institution->banner_inst_picture_5))){
                 unlink(public_path($institution->banner_inst_picture_5));
             }
@@ -486,62 +484,61 @@ class InstitutionController extends Controller
         $input = $request->all();
 
         //Delete old images if necessary
-        if(isset($request->institution_bg_picture)) {
-            if(file_exists(public_path($institution->institution_bg_picture))){
+        if(isset($request->institution_bg_picture) and !empty($institution->institution_bg_picture)) {
+            if(file_exists(public_path($institution->institution_bg_picture)))
                 unlink(public_path($institution->institution_bg_picture));
-            }
         }
-        if(isset($request->institution_picture_1)) {
+        if(isset($request->institution_picture_1) and !empty($institution->institution_picture_1)) {
             if(file_exists(public_path($institution->institution_picture_1))){
                 unlink(public_path($institution->institution_picture_1));
             }
         }
-        if(isset($request->institution_picture_2)) {
+        if(isset($request->institution_picture_2) and !empty($institution->institution_picture_2)) {
             if(file_exists(public_path($institution->institution_picture_2))){
                 unlink(public_path($institution->institution_picture_2));
             }
         }
-        if(isset($request->institution_picture_3)) {
+        if(isset($request->institution_picture_3) and !empty($institution->institution_picture_3)) {
             if(file_exists(public_path($institution->institution_picture_3))){
                 unlink(public_path($institution->institution_picture_3));
             }
         }
-        if(isset($request->institution_picture_4)) {
+        if(isset($request->institution_picture_4) and !empty($institution->institution_picture_4)) {
             if(file_exists(public_path($institution->institution_picture_4))){
                 unlink(public_path($institution->institution_picture_4));
             }
         }
-        if(isset($request->institution_picture_5)) {
+        if(isset($request->institution_picture_5) and !empty($institution->institution_picture_5)) {
             if(file_exists(public_path($institution->institution_picture_5))){
                 unlink(public_path($institution->institution_picture_5));
             }
         }
-        if(isset($request->institution_picture_6)) {
+        if(isset($request->institution_picture_6) and !empty($institution->institution_picture_6)) {
             if(file_exists(public_path($institution->institution_picture_6))){
                 unlink(public_path($institution->institution_picture_6));
             }
         }
-        if(isset($request->banner_inst_picture_1)) {
+        if(isset($request->banner_inst_picture_1) and !empty($institution->banner_inst_picture_1)) {
             if(file_exists(public_path($institution->banner_inst_picture_1))){
                 unlink(public_path($institution->banner_inst_picture_1));
             }
         }
-        if(isset($request->banner_inst_picture_2)) {
+        if(isset($request->banner_inst_picture_2) and !empty($institution->banner_inst_picture_2)) {
             if(file_exists(public_path($institution->banner_inst_picture_2))){
                 unlink(public_path($institution->banner_inst_picture_2));
             }
         }
-        if(isset($request->banner_inst_picture_3)) {
+        if(isset($request->banner_inst_picture_3) and !empty($institution->banner_inst_picture_3)) {
             if(file_exists(public_path($institution->banner_inst_picture_3))){
                 unlink(public_path($institution->banner_inst_picture_3));
             }
         }
-        if(isset($request->banner_inst_picture_4)) {
+        if(isset($request->banner_inst_picture_4) and !empty($institution->banner_inst_picture_4)) {
             if(file_exists(public_path($institution->banner_inst_picture_4))){
                 unlink(public_path($institution->banner_inst_picture_4));
             }
         }
-        if(isset($request->banner_inst_picture_5)) {
+        if(isset($request->banner_inst_picture_5) and !empty($institution->banner_inst_picture_5)) {
             if(file_exists(public_path($institution->banner_inst_picture_5))){
                 unlink(public_path($institution->banner_inst_picture_5));
             }
